cleanup and readme

+++ scripts/development/m3_prom_remote_stack/stop_m3.sh
@@ -2,4 +2,8 @@
 
 set -xe
 
-docker-compose -f docker-compose.yml down
+if [[ "$WIPE_DATA" == true ]]; then
+  docker-compose -f docker-compose.yml down
+else
+  docker-compose -f docker-compose.yml stop
+fi
